refactor(spa): extract quick-command-bindings module + split store tests (#679)
Closes #679 (Q2 + Q3) — codex round-2 followups deferred from PR #677.

## Q2: Pure data module (lib/quick-command-bindings.ts)

`useQuickCommandStore.ts` had grown from 91 → ~225 lines after Phase 1a,
mixing schema (QuickCommand / Bindings types), sanitizer (UNSAFE_KEYS,
sanitizeBindings, getBindingTargets), persist merge function, and the
zustand store + actions + persist + sync glue.

This commit extracts the pure-data primitives — schema, UNSAFE_KEYS,
sanitizeBindings, getBindingTargets, mergePersistedQuickCommandState,
plus a new `QuickCommandData` shape — into the new module. The store
imports from there and is reduced to: state + actions + persist + sync
glue. The sync contributor (and its test) likewise import sanitizeBindings
+ QuickCommand directly from the pure module instead of pulling them
through the store. mergePersistedQuickCommandState is now generic over
`T extends QuickCommandData` so the store can pass its full state and
recover the same shape (action methods preserved through merge).

## Q3: Split store tests into three files

`useQuickCommandStore.test.ts` was 366+ lines; Phase 1b would push it
past 500 once Settings dialog and multi-select chip tests land. Split:

- `lib/quick-command-bindings.test.ts` — sanitizer (forward-compat /
  prototype rejection / non-object payloads), getBindingTargets
  own-property guard (incl. inherited Object.prototype keys),
  mergePersistedQuickCommandState hydrate trust boundary (incl.
  generic-over-T preservation of caller fields).
- `stores/useQuickCommandStore.crud.test.ts` — capability CRUD
  (addCommand / updateCommand / removeCommand / getCommands / per-host
  overrides).
- `stores/useQuickCommandStore.bindings.test.ts` — store-level binding
  integration: setBinding / getBoundCommands (capability order,
  dangling filter, multi-slot mount), prototype-key safety integration,
  null-hostId path (spec §4.4).

53 tests across the four files pass; lint clean; build clean.
Pre-existing failures (`hosts.test.ts` × 3 + `TabBar.test.tsx` × 1)
untouched.
