+++ spa/src/stores/useQuickCommandStore.ts
@@ -2,23 +2,14 @@ import { create } from 'zustand'
 import { persist } from 'zustand/middleware'
 import { purdexStorage, STORAGE_KEYS, syncManager } from '../lib/storage'
 import type { QuickCommandSlotId } from '../lib/quick-command-slots'
+import {
+  type QuickCommand,
+  type QuickCommandData,
+  getBindingTargets,
+  mergePersistedQuickCommandState,
+} from '../lib/quick-command-bindings'
 
-export interface QuickCommand {
-  id: string
-  name: string
-  command: string
-  icon?: string
-  category?: string
-  hostOnly?: boolean
-}
-
-export type Bindings = Record<string /* commandId */, QuickCommandSlotId[]>
-
-interface QuickCommandState {
-  global: QuickCommand[]
-  byHost: Record<string, QuickCommand[]>
-  bindings: Bindings
-
+interface QuickCommandState extends QuickCommandData {
   addCommand: (cmd: QuickCommand, hostId?: string) => void
   updateCommand: (id: string, patch: Partial<QuickCommand>, hostId?: string) => void
   removeCommand: (id: string, hostId?: string) => void
@@ -35,83 +26,6 @@ interface QuickCommandState {
 // (`start-cc` / `start-codex`) keep them but bindings = {} so nothing renders
 // until the user mounts them via Settings.
 const DEFAULT_COMMANDS: QuickCommand[] = []
-
-const UNSAFE_KEYS = new Set(['__proto__', 'constructor', 'prototype'])
-
-/**
- * AR-1 (mirrors `useModuleEnabledStore` codex review #617):
- * Sanitize bindings on rehydrate / sync deserialize. A corrupted payload —
- * hand-edited localStorage, failed migration, hostile sync source — must
- * NEVER let arbitrary string keys silently mount commands into slots.
- *
- * Rules (spec §2.3 — forward-compat with future slot ids):
- *  - Top-level value must be a plain object (rejects arrays, null, primitives).
- *  - Reject `__proto__` / `constructor` / `prototype` command id keys
- *    (prototype pollution).
- *  - Drop entries whose value is not an array.
- *  - Within each array, keep only non-empty strings. NO slot id whitelist —
- *    spec §2.3 explicitly requires accepting unknown slot ids in Phase 1
- *    so cross-version sync (older client pulls newer client's future slot
- *    binding) doesn't lose data; SlotHost simply ignores unknown slot ids
- *    at render time.
- *  - Drop entries whose cleaned array is empty (matches setBinding(id, [])).
- */
-export function sanitizeBindings(raw: unknown): Bindings {
-  if (raw === null || typeof raw !== 'object' || Array.isArray(raw)) return {}
-  const out: Bindings = {}
-  for (const [cmdId, targets] of Object.entries(raw as Record<string, unknown>)) {
-    if (typeof cmdId !== 'string' || cmdId.length === 0) continue
-    if (UNSAFE_KEYS.has(cmdId)) continue
-    if (!Array.isArray(targets)) continue
-    const cleaned: QuickCommandSlotId[] = []
-    for (const t of targets) {
-      if (typeof t === 'string' && t.length > 0) cleaned.push(t as QuickCommandSlotId)
-    }
-    if (cleaned.length > 0) out[cmdId] = cleaned
-  }
-  return out
-}
-
-/**
- * Read bindings[cmdId] safely. Prevents inherited-prop attacks where cmdId
- * happens to match Object.prototype methods (toString / valueOf / hasOwnProperty
- * / isPrototypeOf / etc.) — naive `bindings[cmdId]` would resolve to a
- * non-array function and `.includes(slot)` would throw, crashing
- * getBoundCommands for every caller (DoS).
- *
- * Returns undefined unless `bindings` has its OWN property `cmdId` AND that
- * value is an array.
- */
-function getBindingTargets(bindings: Bindings, cmdId: string): QuickCommandSlotId[] | undefined {
-  if (!Object.prototype.hasOwnProperty.call(bindings, cmdId)) return undefined
-  const targets = bindings[cmdId]
-  return Array.isArray(targets) ? targets : undefined
-}
-
-/**
- * Pure merge function — used by zustand persist `merge` hook AND directly by
- * tests to drive the real hydrate trust boundary with malformed payloads.
- * Exporting this avoids the previous test pattern of using setState as a
- * mock for hydrate, which never actually went through sanitizer.
- */
-export function mergePersistedQuickCommandState(
-  persisted: unknown,
-  current: QuickCommandState,
-): QuickCommandState {
-  const p = persisted as
-    | { global?: unknown; byHost?: unknown; bindings?: unknown }
-    | null
-    | undefined
-  return {
-    ...current,
-    global: Array.isArray(p?.global) ? (p?.global as QuickCommand[]) : current.global,
-    byHost:
-      typeof p?.byHost === 'object' && p?.byHost !== null && !Array.isArray(p?.byHost)
-        ? (p?.byHost as Record<string, QuickCommand[]>)
-        : current.byHost,
-    bindings: sanitizeBindings(p?.bindings),
-  }
-}
 
 export const useQuickCommandStore = create<QuickCommandState>()(
   persist(
